requirements.txt
pycryptodome==3.14.1
Pillow==9.0.1